pyproject.toml
[tool.poetry]
name = "mdexport"
version = "0.5.0"
description = "CLI tool to export markdown to pdf using jinja2 templates."
authors = ["andyverstraeten <andy.verstraeten@gmail.com>"]
readme = "README.md"
packages = [{ include = "mdexport"}] 

[tool.poetry.dependencies]
python = ">=3.10,<4.0"
click = "^8.1.7"
markdown2 = "^2.5.1"
weasyprint = "^63.0"
jinja2 = "^3.1.4"
python-frontmatter = "^1.1.0"
beautifulsoup4 = "^4.12.3"


[tool.poetry.scripts]
mdexport = "mdexport.mdexport:cli"

[tool.poetry.group.dev.dependencies]
pytest = "^8.3.3"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
